Relocate JSON parsing to sub-package (update references)
