+++ src/main/java/io/xlate/edi/internal/stream/json/StaEDIJavaxJsonParser.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  ******************************************************************************/
-package io.xlate.edi.internal.stream;
+package io.xlate.edi.internal.stream.json;
 
 import java.util.Map;
